requirements.txt
fastapi
uvicorn
motor
pymongo
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
python-dotenv
opencv-python
mediapipe
numpy
scikit-learn
razorpay
